+++ packages/@lazy-random/df-array/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.2 (2021-12-12)
+
+
+### 📦　Code Refactoring
+
+* `@lazy-random/df-array` ([6910a7b](https://github.com/bluelovers/ws-random/commit/6910a7b80c94ca50d3daa928e895b69f01d61342))
+
+
+### 🔖　Miscellaneous
+
+* . ([8d815a9](https://github.com/bluelovers/ws-random/commit/8d815a9451f12cabc9b81680e463d429c45f2506))
+* . ([830c654](https://github.com/bluelovers/ws-random/commit/830c6544fcb622f95c789f84d62011165e6990df))
+
+
+
+
+
 ## 1.0.1 (2021-12-12)